+++ packages/dash-table/demo/AppMode.ts
@@ -14,6 +14,8 @@ import {
 import { TooltipSyntax } from 'dash-table/tooltips/props';
 
 export enum AppMode {
+    Clearable = 'clearable',
+    ClearableMerged = 'clearableMerged',
     Date = 'date',
     Default = 'default',
     Filtering = 'filtering',
@@ -185,6 +187,24 @@ function getTypedState() {
     return state;
 }
 
+function getClearableState() {
+    const state = getDefaultState();
+    state.tableProps.filter_action = TableAction.Native;
+
+    R.forEach(c => {
+        c.clearable = true;
+    }, state.tableProps.columns || []);
+
+    return state;
+}
+
+function getClearableMergedState() {
+    const state = getClearableState();
+    state.tableProps.merge_duplicate_headers = true;
+
+    return state;
+}
+
 function getDateState() {
     const state = getTypedState();
 
@@ -325,6 +345,10 @@ function getState() {
     const mode = Environment.searchParams.get('mode');
 
     switch (mode) {
+        case AppMode.Clearable:
+            return getClearableState();
+        case AppMode.ClearableMerged:
+            return getClearableMergedState();
         case AppMode.Date:
             return getDateState();
         case AppMode.Filtering: