+++ packages/dash-table/src/dash-table/utils/actions.js
@@ -5,6 +5,7 @@ function getGroupedColumnIndices(column, columns, headerRowIndex, mergeDuplicate
     if (!column.name || (Array.isArray(column.name) && column.name.length < headerRowIndex) || !mergeDuplicateHeaders) {
         return { groupIndexFirst: columnIndex, groupIndexLast: columnIndex };
     }
+
     let lastColumnIndex = columnIndex;
 
     for (let i = columnIndex; i < columns.length; ++i) {
@@ -16,25 +17,42 @@ function getGroupedColumnIndices(column, columns, headerRowIndex, mergeDuplicate
             break;
         }
     }
+
     return { groupIndexFirst: columnIndex, groupIndexLast: lastColumnIndex };
 }
 
-export function deleteColumn(column, columns, headerRowIndex, data) {
-    const {groupIndexFirst, groupIndexLast} = getGroupedColumnIndices(
-        column, columns, headerRowIndex
+export function getAffectedColumns(column, columns, headerRowIndex, mergeDuplicateHeaders) {
+    const { groupIndexFirst, groupIndexLast } = getGroupedColumnIndices(
+        column, columns, headerRowIndex, mergeDuplicateHeaders, columns.indexOf(column)
     );
-    const rejectedColumnIds = R.slice(
+
+    return R.slice(
         groupIndexFirst,
         groupIndexLast + 1,
         R.pluck('id', columns)
     );
+}
+
+export function clearColumn(column, columns, headerRowIndex, mergeDuplicateHeaders, data) {
+    const rejectedColumnIds = getAffectedColumns(column, columns, headerRowIndex, mergeDuplicateHeaders);
+
+    return {
+        data: R.map(R.omit(rejectedColumnIds), data)
+    };
+}
+
+export function deleteColumn(column, columns, headerRowIndex, mergeDuplicateHeaders, data) {
+    const {groupIndexFirst, groupIndexLast} = getGroupedColumnIndices(
+        column, columns, headerRowIndex, mergeDuplicateHeaders, columns.indexOf(column)
+    );
+
     return {
         columns: R.remove(
             groupIndexFirst,
             1 + groupIndexLast - groupIndexFirst,
             columns
         ),
-        data: R.map(R.omit(rejectedColumnIds), data),
+        ...clearColumn(column, columns, headerRowIndex, mergeDuplicateHeaders, data),
         // NOTE - We're just clearing these so that there aren't any
         // inconsistencies. In an ideal world, we would probably only
         // update them if they contained one of the columns that we're