+++ packages/dash-table/tests/cypress/src/DashTable.ts
@@ -3,26 +3,42 @@ export default class DashTable {
         return cy.get(`#table tbody tr td.column-${column}`).eq(row);
     }
 
-    static focusCell(row: number, column: number) {
-        // somehow we need to scrollIntoView AFTER click, or it doesn't
-        // work right. Why?
-        return this.getCell(row, column).click().scrollIntoView();
-    }
-
-    static getCellById(row: number, column: number | string) {
+    static getCellById(row: number, column: string) {
         return cy.get(`#table tbody tr td[data-dash-column="${column}"]`).eq(row);
-    }
-
-    static focusCellById(row: number, column: number | string) {
-        return this.getCellById(row, column).click().scrollIntoView();
     }
 
     static getFilter(column: number) {
         return cy.get(`#table tbody tr th.dash-filter.column-${column}`);
     }
 
-    static getFilterById(column: number | string) {
+    static getFilterById(column: string) {
         return cy.get(`#table tbody tr th.dash-filter[data-dash-column="${column}"]`);
+    }
+
+    static getHeader(row: number, column: number) {
+        return cy.get(`#table tbody tr th.dash-header.column-${column}`).eq(row);
+    }
+
+    static getHeaderById(row: number, column: string) {
+        return cy.get(`#table tbody tr th.dash-header[data-dash-column="${column}"]`).eq(row);
+    }
+
+    static focusCell(row: number, column: number) {
+        // somehow we need to scrollIntoView AFTER click, or it doesn't
+        // work right. Why?
+        return this.getCell(row, column).click().scrollIntoView();
+    }
+
+    static focusCellById(row: number, column: string) {
+        return this.getCellById(row, column).click().scrollIntoView();
+    }
+
+    static clearColumnById(row: number, column: string) {
+        return cy.get(`#table tbody tr th.dash-header[data-dash-column="${column}"] .column-header--clear`).eq(row).click();
+    }
+
+    static deleteColumnById(row: number, column: string) {
+        return cy.get(`#table tbody tr th.dash-header[data-dash-column="${column}"] .column-header--delete`).eq(row).click();
     }
 
     static getDelete(row: number) {